+++ src/js/components/FormField/FormField.js
@@ -28,7 +28,7 @@ const grommetInputNames = [
   'TextInput',
   'Select',
   'MaskedInput',
-  'MultiSelect',
+  'SelectMultiple',
   'TextArea',
   'DateInput',
   'FileInput',
