--- src/js/components/MultiSelect/README.md
@@ -1,2 +1,0 @@
-## MultiSelect
-Documentation for this component: https://v2.grommet.io/multiselect
