--- src/js/components/MultiSelect/index.js
@@ -1,1 +1,0 @@
-export { MultiSelect } from './MultiSelect';
