+++ src/js/components/SelectMultiple/README.md
@@ -1,0 +1,2 @@
+## SelectMultiple
+Documentation for this component: https://v2.grommet.io/SelectMultiple
