+++ src/js/components/SelectMultiple/index.d.ts
@@ -1,19 +1,20 @@
 import * as React from 'react';
 import { BasicSelectProps } from '../Select';
 
-export interface MultiSelectProps extends BasicSelectProps {
+export interface SelectMultipleProps extends BasicSelectProps {
   defaultValue?: (string | number | object)[];
   helpContent?: React.ReactNode;
+  limit?: number;
   value?: (string | number | object)[];
   showSelectedInline?: boolean;
 }
 
 // Try without Omit<> to see where we define our own attributes for overrides
 // value, name, id, onChange, placeholder
-export interface MultiSelectExtendedProps
-  extends MultiSelectProps,
-    Omit<JSX.IntrinsicElements['input'], keyof MultiSelectProps> {}
+export interface SelectMultipleExtendedProps
+  extends SelectMultipleProps,
+    Omit<JSX.IntrinsicElements['input'], keyof SelectMultipleProps> {}
 
-declare const MultiSelect: React.FC<MultiSelectExtendedProps>;
+declare const SelectMultiple: React.FC<SelectMultipleExtendedProps>;
 
-export { MultiSelect };
+export { SelectMultiple };
