+++ src/js/components/SelectMultiple/index.js
@@ -1,0 +1,1 @@
+export { SelectMultiple } from './SelectMultiple';
