+++ src/js/components/index.d.ts
@@ -40,7 +40,6 @@ export * from './Markdown';
 export * from './MaskedInput';
 export * from './Menu';
 export * from './Meter';
-export * from './MultiSelect';
 export * from './PageHeader';
 export * from './Pagination';
 export * from './Paragraph';
@@ -55,6 +54,7 @@ export * from './RangeSelector';
 export * from './RoutedAnchor';
 export * from './RoutedButton';
 export * from './Select';
+export * from './SelectMultiple';
 export * from './Sidebar';
 export * from './SkipLink';
 export * from './SkipLinks';
