+++ src/js/components/index.js
@@ -40,7 +40,6 @@ export * from './Markdown';
 export * from './MaskedInput';
 export * from './Menu';
 export * from './Meter';
-export * from './MultiSelect';
 export * from './Page';
 export * from './PageContent';
 export * from './PageHeader';
@@ -57,6 +56,7 @@ export * from './RangeSelector';
 export * from './RoutedAnchor';
 export * from './RoutedButton';
 export * from './Select';
+export * from './SelectMultiple';
 export * from './Sidebar';
 export * from './SkipLink';
 export * from './SkipLinks';
