+++ src/js/index.d.ts
@@ -40,7 +40,6 @@ export * from './components/Markdown';
 export * from './components/MaskedInput';
 export * from './components/Menu';
 export * from './components/Meter';
-export * from './components/MultiSelect';
 export * from './components/Page';
 export * from './components/PageContent';
 export * from './components/PageHeader';
@@ -57,6 +56,7 @@ export * from './components/RangeSelector';
 export * from './components/RoutedAnchor';
 export * from './components/RoutedButton';
 export * from './components/Select';
+export * from './components/SelectMultiple';
 export * from './components/Sidebar';
 export * from './components/SkipLink';
 export * from './components/SkipLinks';
